Struct sprite::SpriteMap

pub struct SpriteMap {
    pub width: u32,
    pub height: u32,
    pub regions: HashMap<String, SpriteRegion>,
    pub url: String,
}

The sprite map holding all the images for a given invocation.

Fields

width

height

regions

url

Methods

impl SpriteMap

fn build(glob: &str, output: &Path, base_url: &str) -> SpriteMap

Generates a css sprite map from the files matching the glob pattern. Notes: - the glob parameter is a folder for the time being - only PNG files can be made into css sprites at this time.

fn region(&self, name: &str) -> Option<&SpriteRegion>

Return the region with the given name (compass API).

fn names(&self) -> Vec<&str>

Return the names of all the regions (compass API).

fn css_background(&self, region: &SpriteRegion, offset_x: i32, offset_y: i32) -> String

Returns the image and background position for use in a single shorthand property. For example: background: spriteMap.sprite("name") could generate: url('/images/icons.png') 0 -24px no-repeat;